refactor: major cleanup of file structure
