+++ frontend/src/lib/components/editors/DocumentEditor.svelte
