+++ frontend/src/lib/components/elements/LoadingIcon.svelte
