+++ frontend/src/lib/components/sidebar/MockDirectory.svelte
