+++ frontend/src/lib/components/sidebar/SideBar.svelte
