+++ frontend/src/lib/components/topbar/TopBar.svelte
