+++ frontend/src/lib/types.ts
@@ -1,10 +1,19 @@
-// TODO: rename this file to remove the .d, it's bad
+/**
+ * @file
+ * Misc type declarations
+ */
 
+/**
+ * All configurable permissions
+ */
 export enum Permission {
 	ManageUsers = 'ManageUsers',
 	ManageContent = 'ManageContent'
 }
 
+/**
+ * A map between the internal permission representations and the "pretty print" representation
+ */
 export const allPermissions: Map<Permission, string> = new Map();
 allPermissions.set(Permission.ManageContent, 'Manage Content');
 allPermissions.set(Permission.ManageUsers, 'Manage Users');
